devlink: print VFIO device name if available
If the kernel provides the DEVLINK_ATTR_VFIODEV_NAME attribute, print the
device name when displaying port information. Since this is the name of the
struct device, just use "device" rather than vfio_device for the shorthand.
Since the type is vfio already and the device name contains vfio it is
obvious that this is a vfio device.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>